zoo/legacy_3_ROSA_Copilot/ROSA_UNIX/summer2023/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
boto3==1.27.0
botocore==1.30.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.4
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.4
comm==0.1.3
contourpy==1.1.0
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
distlib==0.3.6
executing==1.2.0
filelock==3.12.2
Fiona==1.9.4.post1
fonttools==4.40.0
frozenlist==1.3.3
geopandas==0.13.2
huggingface-hub==0.11.1
idna==3.4
importlib-metadata==6.7.0
importlib-resources==5.12.0
intel-openmp==2023.2.0
ipykernel==6.24.0
ipython==8.12.2
jedi==0.18.2
jmespath==1.0.1
jupyter_client==8.3.0
jupyter_core==5.3.1
kiwisolver==1.4.4
matplotlib==3.7.1
matplotlib-inline==0.1.6
matplotlib-scalebar==0.8.1
mkl==2023.2.0
mkl-service==2.4.0
multidict==6.0.4
nest-asyncio==1.5.6
numpy==1.24.4
openai==0.27.8
packaging==23.1
pandas==2.0.3
parso==0.8.3
pickleshare==0.7.5
Pillow==10.0.0
pipenv==2023.7.4
platformdirs==3.8.0
prompt-toolkit==3.0.38
psutil==5.9.5
pure-eval==0.2.2
Pygments==2.15.1
pyparsing==3.1.0
pyproj==3.5.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pywin32==306
PyYAML==6.0.1
pyzmq==25.1.0
qrcode==6.1
regex==2023.6.3
requests==2.31.0
s3transfer==0.6.1
schedule==1.2.0
seaborn==0.12.2
shapely==2.0.1
six==1.15.0
stack-data==0.6.2
tbb==2021.10.0
tokenizers==0.13.2
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.25.1
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.25.11
virtualenv==20.23.1
virtualenv-clone==0.5.7
wcwidth==0.2.6
yarl==1.9.2
zipp==3.15.0
